+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 - Tests should now leverage default CloudHarvestStack configurations instead of internally managed `tests/docker-compose.yaml`
 - `task_chain_from_dict()` now merges `**kwargs` with the task chain configuration on class instantiation
 - Removed the `HeartBeat` class. Moved `silos.py` into the root package directory
-- Added `HarvestRecord.assign_value_to_key()`
+- Added methods to `HarvestRecord`
+  - `assign_value_to_key()`
+  - `title_keys()`
 
 # 0.4.2
 - Added `BaseHarvestTaskChain` and `HarvestRecordUpdateTask` to upload data collected from other sources to a persistent silo